Add files via upload
In my role as a trainer and moderator at IcodeGuru, I deliver comprehensive lectures for the "Data Analysis with Python" course to a diverse audience comprising both graduate and undergraduate students. I conduct these one-hour sessions regularly at 7 am in the local language, ensuring accessibility and effective communication to facilitate a deeper understanding of the subject matter. My commitment lies in empowering students with practical skills and insights that are essential in the dynamic field of data analysis.